utils/generator-validator/requirements.txt
pyshark==0.4.2.9
PyYAML==5.3
